feat(schema)!: widen identity columns to bigint/bigserial
All identity PKs and their FKs (users, tenants, roles, user_roles,
auth_domains, communication_channels, user_communication_methods, sessions,
dev_otp_enrollments) widen serial/integer -> bigserial/bigint, and
current_user_id()/resolve_session() return bigint. Matches the bigint audit
columns from @smplcty/schema-std and lets bigint apps consume the schema by
reference without a type conflict (schema-flow extend: can't change an
imported column's type).

The TS API is unchanged (ids stay `number`): the library now parses the
bigint values pg returns as strings back to number at every read boundary
(createSession, withSession, validateSession, findUserByCommunicationMethod,
withServiceContext, the method router, and the http handlers). Adds a test
proving an id beyond int4 range round-trips.

BREAKING CHANGE: a database already created with the previous int4 schema
must widen its identity columns via a one-off pre-script (schema-flow can't
do serial->bigserial declaratively). Fresh installs are unaffected.

+++ src/scope/flat-tenant.ts
@@ -25,7 +25,7 @@ const SELECT_TENANTS = `
   SELECT
     COALESCE(
       array_agg(DISTINCT tenant_id) FILTER (WHERE tenant_id IS NOT NULL),
-      '{}'::integer[]
+      '{}'::bigint[]
     ) AS tenant_ids,
     COALESCE(bool_or(tenant_id IS NULL), false) AS all_tenants
   FROM user_roles
